+++ ressched/services/resschedmgr/resschedfwk/src/res_sched_mgr.cpp
@@ -70,7 +70,7 @@ void ResSchedMgr::DispatchResourceInner(uint32_t resType, int64_t value, const J
 extern "C" void ReportDataInProcess(uint32_t resType, int64_t value, const Json::Value& payload)
 {
     RESSCHED_LOGI("ResSchedMgr::ReportDataInProcess receive resType = %{public}u, value = %{public}lld.",
-        resType, value);
+        resType, (long long)value);
     ResSchedMgr::GetInstance().ReportData(resType, value, payload);
 }
 } // namespace ResourceSchedule
